# src/lib/swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OrbitABM API",
    "description": "RESTful API for OrbitABM - Account-Based Marketing Campaign Intelligence Platform",
    "version": "1.0.0",
    "contact": {
      "name": "OrbitABM Support",
      "email": "support@orbitabm.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://orbitabm.com/license"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    },
    {
      "url": "https://your-domain.com/api",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/campaigns": {
      "get": {
        "tags": ["Campaigns"],
        "summary": "List campaigns",
        "description": "Retrieve a list of campaigns with optional filtering and pagination",
        "parameters": [
          {
            "name": "organization_id",
            "in": "query",
            "required": true,
            "description": "Organization UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by campaign status",
            "schema": {
              "type": "string",
              "enum": ["planned", "active", "paused", "completed", "won", "lost", "pivoted"]
            }
          },
          {
            "name": "market_id",
            "in": "query",
            "description": "Filter by market UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "vertical_id",
            "in": "query",
            "description": "Filter by vertical UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "assigned_to",
            "in": "query",
            "description": "Filter by assigned user UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "company_id",
            "in": "query",
            "description": "Filter by company UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of results per page",
            "schema": {
              "type": "integer",
              "default": 100,
              "minimum": 1,
              "maximum": 1000
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of results to skip",
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/CampaignWithRelations"
                      }
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "count": {
                      "type": "integer",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "post": {
        "tags": ["Campaigns"],
        "summary": "Create campaign",
        "description": "Create a new campaign",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CampaignCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Campaign created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/CampaignWithRelations"
                    },
                    "success": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Campaign": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Campaign UUID"
          },
          "organization_id": {
            "type": "string",
            "format": "uuid",
            "description": "Organization UUID"
          },
          "name": {
            "type": "string",
            "description": "Campaign name",
            "example": "Acme HVAC Q1 Campaign"
          },
          "company_id": {
            "type": "string",
            "format": "uuid",
            "description": "Target company UUID"
          },
          "playbook_template_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "Playbook template UUID"
          },
          "market_id": {
            "type": "string",
            "format": "uuid",
            "description": "Market UUID"
          },
          "vertical_id": {
            "type": "string",
            "format": "uuid",
            "description": "Vertical UUID"
          },
          "status": {
            "type": "string",
            "enum": ["planned", "active", "paused", "completed", "won", "lost", "pivoted"],
            "description": "Campaign status",
            "example": "active"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "description": "Campaign start date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "nullable": true,
            "description": "Campaign end date"
          },
          "current_step": {
            "type": "integer",
            "nullable": true,
            "description": "Current playbook step",
            "example": 3
          },
          "pivot_reason": {
            "type": "string",
            "nullable": true,
            "description": "Reason for pivot"
          },
          "pivot_to_campaign_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "Pivot target campaign UUID"
          },
          "assigned_to": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "Assigned user UUID"
          },
          "notes": {
            "type": "string",
            "nullable": true,
            "description": "Additional notes"
          },
          "value_proposition": {
            "type": "string",
            "nullable": true,
            "description": "Value proposition for the campaign"
          },
          "primary_wedge": {
            "type": "string",
            "nullable": true,
            "description": "Primary wedge/entry point"
          },
          "backup_trigger": {
            "type": "string",
            "nullable": true,
            "description": "Backup trigger for follow-up"
          },
          "success_criteria": {
            "type": "string",
            "nullable": true,
            "description": "Success criteria for the campaign"
          },
          "research_doc_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "Research document UUID"
          },
          "sequence_doc_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "description": "Sequence document UUID"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Deletion timestamp"
          }
        },
        "required": ["id", "organization_id", "name", "company_id", "market_id", "vertical_id", "status"]
      },
      "CampaignWithRelations": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Campaign"
          },
          {
            "type": "object",
            "properties": {
              "companies": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "name": {
                    "type": "string",
                    "example": "Acme HVAC"
                  },
                  "website": {
                    "type": "string",
                    "format": "uri",
                    "example": "https://acmehvac.com"
                  },
                  "estimated_revenue": {
                    "type": "number",
                    "example": 2500000
                  }
                }
              },
              "markets": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "name": {
                    "type": "string",
                    "example": "Fort Wayne, IN"
                  },
                  "state": {
                    "type": "string",
                    "example": "IN"
                  }
                }
              },
              "verticals": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "name": {
                    "type": "string",
                    "example": "HVAC Companies"
                  },
                  "sector": {
                    "type": "string",
                    "example": "Home Services"
                  }
                }
              },
              "playbook_templates": {
                "type": "object",
                "nullable": true,
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "name": {
                    "type": "string",
                    "example": "Standard HVAC Outreach"
                  },
                  "total_duration_days": {
                    "type": "integer",
                    "example": 45
                  }
                }
              },
              "profiles": {
                "type": "object",
                "nullable": true,
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "full_name": {
                    "type": "string",
                    "example": "John Smith"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "john@company.com"
                  }
                }
              }
            }
          }
        ]
      },
...
      "CampaignCreate": {
        "type": "object",
        "properties": {
          "organization_id": {
            "type": "string",
            "format": "uuid",
            "description": "Organization UUID"
          },
          "name": {
            "type": "string",
            "description": "Campaign name",
            "example": "Acme HVAC Q1 Campaign"
          },
          "company_id": {
            "type": "string",
            "format": "uuid",
            "description": "Target company UUID"
          },
          "playbook_template_id": {
            "type": "string",
            "format": "uuid",
            "description": "Playbook template UUID"
          },
          "market_id": {
            "type": "string",
            "format": "uuid",
            "description": "Market UUID"
          },
          "vertical_id": {
            "type": "string",
            "format": "uuid",
            "description": "Vertical UUID"
          },
          "status": {
            "type": "string",
            "enum": ["planned", "active", "paused", "completed", "won", "lost", "pivoted"],
            "default": "planned",
            "description": "Campaign status"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "description": "Campaign start date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "description": "Campaign end date"
          },
          "current_step": {
            "type": "integer",
            "default": 0,
            "description": "Current playbook step"
          },
          "assigned_to": {
            "type": "string",
            "format": "uuid",
            "description": "Assigned user UUID"
          },
          "notes": {
            "type": "string",
            "description": "Additional notes"
          },
          "value_proposition": {
            "type": "string",
            "description": "Value proposition for the campaign"
          },
          "primary_wedge": {
            "type": "string",
            "description": "Primary wedge/entry point"
          },
          "backup_trigger": {
            "type": "string",
            "description": "Backup trigger for follow-up"
          },
          "success_criteria": {
            "type": "string",
            "description": "Success criteria for the campaign"
          },
          "research_doc_id": {
            "type": "string",
            "format": "uuid",
            "description": "Research document UUID"
          },
          "sequence_doc_id": {
            "type": "string",
            "format": "uuid",
            "description": "Sequence document UUID"
          }
        },
        "required": ["organization_id", "name", "company_id", "market_id", "vertical_id"]
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Human-readable error message",
            "example": "Company with this name already exists"
          },
          "code": {
            "type": "string",
            "description": "Error code",
            "enum": ["VALIDATION_ERROR", "NOT_FOUND", "DUPLICATE_ENTRY", "FORBIDDEN", "DATABASE_ERROR", "SERVER_ERROR", "NETWORK_ERROR"],
            "example": "DUPLICATE_ENTRY"
          },
          "success": {
            "type": "boolean",
            "example": false
          },
          "details": {
            "type": "object",
            "nullable": true,
            "description": "Additional error details",
            "example": {
              "field": "name",
              "value": "Acme HVAC"
            }
          }
        },
        "required": ["error", "code", "success"]
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad Request - Invalid input data",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Organization ID is required",
              "code": "VALIDATION_ERROR",
              "success": false
            }
          }
        }
      },
      "NotFound": {
        "description": "Not Found - Resource does not exist",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Company not found",
              "code": "NOT_FOUND",
              "success": false
            }
          }
        }
      },
      "Conflict": {
        "description": "Conflict - Resource already exists or constraint violation",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Company with this name already exists",
              "code": "DUPLICATE_ENTRY",
              "success": false,
              "details": {
                "field": "name",
                "value": "Acme HVAC"
              }
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Internal Server Error - Unexpected server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Internal server error",
              "code": "SERVER_ERROR",
              "success": false
            }
          }
        }
      }
    },
...
  }
}